Fix: Refactor the frontend auth flow to match new preference (#19)
As the user(whether authenticated or not) clicks on get started(either login or signup in the auth part), we set the status to loading and show a launcher page with the logo while a network request is made to check the user's auth status. If the user is not authenticated, he will be sent to the login page, if the user is authenticated, he will be sent to the dashboard page. This logic is also implemented on the protected or dashboard/layout.tsx part.